Add suspense to board subheader
